requirements.txt
streamlit
Pillow
streamlit-3d